Move wallet/internal/compat to internal/compat
This package used to be part of the wallet submodule, but makes more
sense to now be in the top-level internal directory structure after
all submodules were removed.
