+++ internal/compat/compat.go
